refactor: 프로젝트 구조 세부 조정
