+++ src/main/java/http/request/HttpRequestParser.java
@@ -1,4 +1,4 @@
-package util.httprequest;
+package http.request;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
