+++ src/main/java/http/response/ContentType.java
@@ -1,4 +1,4 @@
-package util;
+package http.response;
 
 import java.util.HashMap;
 import java.util.Map;
